* ar.c (print_contents): Cast args of bfd_seek to the right types. 	Replace bfd_read call with call to bfd_bread. 	(extract_file): Likewise.
	* objdump.c (dump_section_header): Cast section->filepos to
	unsigned long to suit printf format string.
	* readelf.c (process_section_headers): Similarly for sh_offset.
	(process_unwind): Likewise.

+++ binutils/objdump.c
@@ -287,7 +287,7 @@ dump_section_header (abfd, section, ignored)
   bfd_printf_vma (abfd, bfd_get_section_vma (abfd, section));
   printf ("  ");
   bfd_printf_vma (abfd, section->lma);
-  printf ("  %08lx  2**%u", section->filepos,
+  printf ("  %08lx  2**%u", (unsigned long) section->filepos,
 	  bfd_get_section_alignment (abfd, section));
   if (! wide_output)
     printf ("\n                ");
